delete 6 months old broken spiders
